+++ src/frontend/src/app/core/state/proposal/index.ts
@@ -1,1 +1,2 @@
 export * from './proposal.model';
+export * from './proposal.service';
